Answering 2 (HARD) should mark card as learning; Relearning cards shouldn't have ivl reset to 1
The common scenario happening is that, I would like to mark card as hard, or as 'almost got it right'
but I wouldn't like it to mark it as 'Bad'. However I would like to repeat the card to fix my mistake.
Then I've changed the meaning of answering 'Hard'. Now it mark card as relearning but don't touch
'ivl', it'll change factor of card (but in a smaller degree than answering 'bad').

Marking card as 'Bad' shouldn't reset ivl to 1. It's unreasonable and inaccurate and it causes
that a lot of card piles up. The reasonable approach is to divide ivl by factor to get previous
ivl. So in fixed spaced repetition is equivalent of moving card to the previous box (not the 1st one).