Move to the previous step instead to the first step, when answered incorrectly
When the word is new or learned, there are a few steps which has to be done
in order to mark word as learned/relearned. However it doesn't make sense
to take a word to the first step in a case of mistake. The same effect will
have moving to the previous step, and it will consume less time to learn
new words.