Remove cards which were answered wrong and for which sessions started previous days
Use cases:
* An user set to big number of reviews of just number of reviews that day is too big, so
an user finish before are cards are moved to "DUE_QUEUE"; than the next day he/she starts
with previous cards, which is demotivating. It's better to move reset those cards
LEARNING to NEW, RELEARNING to DUE and increase lapses as a penalty.
* Usually cards which aren't finished in a given session and are left for other sessions,
are hard cards, or badly written cards - leech cards. They need to be fixed.

Replaced magic number with constants.